[util] Create a utility function for sending commands to the terminal
